Ajustes importe materiales

+++ api/src/routes/app/cost/basic/routeMaterials.php
@@ -87,7 +87,7 @@ $app->post('/materialsDataValidation', function (Request $request, Response $res
 
         if (sizeof($dataImportMaterial) == 0) {
             for ($i = 0; $i < sizeof($materials); $i++) {
-                if (intval($materials[$i]['costRawMaterial']) == false) {
+                if (floatval($materials[$i]['costRawMaterial']) == false) {
                     $i = $i + 2;
                     $dataImportMaterial = array('error' => true, 'message' => "El costo debe ser mayor a cero (0), fila: $i");
                     break;
